Former-commit-id: f47ccf1bab6b4c85570152529a41e885ed97982f

+++ src/com/vimukti/accounter/web/client/core/ClientItem.java
@@ -479,7 +479,7 @@ public class ClientItem extends BaseReport implements IAccounterCore,
 		return onhandQuantity;
 	}
 
-	public void setOnhandQuantity(int onhandQuantity) {
+	public void setOnhandQuantity(long onhandQuantity) {
 		this.onhandQuantity = onhandQuantity;
 	}
